# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Survey Service OpenApi3 Documentation",
        "description": "Behandam Survey Service Api documentation",
        "contact": {
            "email": "admin@kermany.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "LocalHost API Server"
        },
        {
            "url": "https://debug.behaminplus.ir/survey-service",
            "description": "Debug API Server"
        },
        {
            "url": "https://develop.behaminplus.ir/survey-service",
            "description": "DEVELOP API Server"
        }
    ],
    "paths": {
...
        "/api/v1/campaigns/{campaignId}": {
            "get": {
                "tags": [
                    "Campaigns"
                ],
                "summary": "get item by id",
                "operationId": "getCampaignItemById",
                "parameters": [
                    {
                        "name": "X-Proxy-Token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "4fVB9SZidiBAADD2444nLZxxbWk92UcPQkwM8k"
                    },
                    {
                        "name": "campaignId",
                        "in": "path",
                        "description": "ID of campaign",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Campaigns"
                ],
                "summary": "update campaign item by id",
                "operationId": "updateCampaignItemById",
                "parameters": [
                    {
                        "name": "X-Proxy-Token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "4fVB9SZidiBAADD2444nLZxxbWk92UcPQkwM8k"
                    },
                    {
                        "name": "campaignId",
                        "in": "path",
                        "description": "ID of campaign",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CampaignStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Campaigns"
                ],
                "summary": "remove item by id",
                "operationId": "removeAnItemById",
                "parameters": [
                    {
                        "name": "X-Proxy-Token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "4fVB9SZidiBAADD2444nLZxxbWk92UcPQkwM8k"
                    },
                    {
                        "name": "campaignId",
                        "in": "path",
                        "description": "ID of campaign",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
...
        "/api/v1/answers/packages/{packageId}": {
            "get": {
                "tags": [
                    "Answers"
                ],
                "summary": "get list of all asnwer items for this package group by users",
                "operationId": "getListOfAllAnswerItemsForPackage",
                "parameters": [
                    {
                        "name": "X-Proxy-Token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "4fVB9SZidiBAADD2444nLZxxbWk92UcPQkwM8k"
                    },
                    {
                        "name": "packageId",
                        "in": "path",
                        "description": "ID of package",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/api/v1/answers/questions/{questionId}": {
            "get": {
                "tags": [
                    "Answers"
                ],
                "summary": "get answer items by question id",
                "operationId": "getAnswerItemByQuestionId",
                "parameters": [
                    {
                        "name": "X-Proxy-Token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "4fVB9SZidiBAADD2444nLZxxbWk92UcPQkwM8k"
                    },
                    {
                        "name": "questionId",
                        "in": "path",
                        "description": "ID of question",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Answers"
                ],
                "summary": "update Answer item by question id",
                "operationId": "updateAnswerItemByQuestionId",
                "parameters": [
                    {
                        "name": "X-Proxy-Token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "4fVB9SZidiBAADD2444nLZxxbWk92UcPQkwM8k"
                    },
                    {
                        "name": "questionId",
                        "in": "path",
                        "description": "ID of question",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "61"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AnswerStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Answers"
                ],
                "summary": "define a new answer item for this Question by question id",
                "operationId": "postANewAnswerItemForThisQuestion",
                "parameters": [
                    {
                        "name": "X-Proxy-Token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "4fVB9SZidiBAADD2444nLZxxbWk92UcPQkwM8k"
                    },
                    {
                        "name": "questionId",
                        "in": "path",
                        "description": "ID of question",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AnswerStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/api/v1/answers/{answerId}/questions": {
            "get": {
                "tags": [
                    "Answers"
                ],
                "summary": "get answer item by id",
                "operationId": "getAnswerItemById",
                "parameters": [
                    {
                        "name": "X-Proxy-Token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "4fVB9SZidiBAADD2444nLZxxbWk92UcPQkwM8k"
                    },
                    {
                        "name": "answerId",
                        "in": "path",
                        "description": "ID of answer",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/api/v1/answers/{answerId}": {
            "delete": {
                "tags": [
                    "Answers"
                ],
                "summary": "remove and app by id",
                "operationId": "removeAnItemById",
                "parameters": [
                    {
                        "name": "X-Proxy-Token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "4fVB9SZidiBAADD2444nLZxxbWk92UcPQkwM8k"
                    },
                    {
                        "name": "answerId",
                        "in": "path",
                        "description": "ID of answer",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
...
        "/api/v1/packages/{packageId}": {
            "get": {
                "tags": [
                    "Packages"
                ],
                "summary": "get package item by id",
                "operationId": "getPackageItemById",
                "parameters": [
                    {
                        "name": "X-Proxy-Token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "4fVB9SZidiBAADD2444nLZxxbWk92UcPQkwM8k"
                    },
                    {
                        "name": "packageId",
                        "in": "path",
                        "description": "ID of package",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Packages"
                ],
                "summary": "update package item by id",
                "operationId": "updatepackageItemById",
                "parameters": [
                    {
                        "name": "X-Proxy-Token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "4fVB9SZidiBAADD2444nLZxxbWk92UcPQkwM8k"
                    },
                    {
                        "name": "packageId",
                        "in": "path",
                        "description": "ID of package",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PackageRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Packages"
                ],
                "summary": "remove and app by id",
                "operationId": "removeAnItemById",
                "parameters": [
                    {
                        "name": "X-Proxy-Token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "4fVB9SZidiBAADD2444nLZxxbWk92UcPQkwM8k"
                    },
                    {
                        "name": "packageId",
                        "in": "path",
                        "description": "ID of package",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "AnswerStoreRequest": {
                "title": "AnswerStoreRequest",
                "description": "Answer Store Request body data",
                "required": [
                    "package_id",
                    "question_id",
                    "user_id"
                ],
                "type": "object",
                "example": {
                    "title": "YOUR TITLE OF ...",
                    "description": "YOUR TITLE OF ...",
                    "package_id": 1,
                    "question_id": 1,
                    "user_id": 1,
                    "choice_id": 2,
                    "choice_ids": [
                        1,
                        2,
                        3
                    ],
                    "client_app_id": 1
                }
            },
...
            "CampaignStoreRequest": {
                "title": "CampaignStoreRequest",
                "description": "Campaign Store Request body data",
                "required": [
                    "title"
                ],
                "type": "object",
                "example": {
                    "title": "YOUR TITLE OF ..."
                }
            },
...
            "PackageRequest": {
                "title": "PackageRequest",
                "description": "Package Request",
                "required": [
                    "title"
                ],
                "type": "object",
                "example": {
                    "title": "نام پکیج",
                    "package_type_id": 1,
                    "client_app_id": 1,
                    "parent_id": null,
                    "description": "description",
                    "first_text": "first_text",
                    "final_text": "final_text",
                    "started_at": "2021-06-02T14:48:37",
                    "finished_at": "2021-06-02T14:48:37",
                    "is_active": 1,
                    "is_deletable": 1
                }
            },
...
        },
        "securitySchemes": {
            "access-token": {
                "type": "http",
                "description": "Use Custom access token",
                "name": "AccessToken",
                "scheme": "access-token"
            }
        }
    },
...
    "security": [
        []
    ]
}